fix: lint on all files
