+++ index.html
@@ -6,6 +6,6 @@
     <title>Index</title>
   </head>
   <body>
-    <script src="./sample.js"></script>
+    <script src="./src/sample.js"></script>
   </body>
 </html>
